==> src/3.2_run_constraint_scan.ipynb <==
---
output-file: runconstraintscan.html

skip_showdoc: true
---



# Running Constraint Scan

# | default_exp runconstraintscan

This notebook can either be run as a notebook, or using nbdev a python script is produced that can be run in the command line for easier running on e.g. a cluster. It is a notebook in the nbs directory and exported as a .py file in the mmon-gcm library.

# | export

import sys

import cobra
import mmon_gcm.constraintscan  # from x import y syntax doesn't work because of nbdev export format
import pandas as pd
from pandarallel import pandarallel

# | export

test = False

# This cell isn't exported to the .py file, so define here if running in notebook rather than as .py on e.g.a cluster
# This is also where to change whether the light is blue or white if you want to run in the notebook

sys.argv = ["script_name",
            "../outputs/constraint_scan/constraint_scan_results_test.csv",
            "../models/4_stage_GC.json",
            "../outputs/constraint_scan/constraints_df.csv",
            "white",
            "2"]

# | export

results_path = sys.argv[1]
model_path = sys.argv[2]
constraints_csv = sys.argv[3]
light_colour = sys.argv[4]
no_cores = int(sys.argv[5])

# | export
constraints = pd.read_csv(constraints_csv, index_col=[0], header=[0])
model = cobra.io.load_json_model(model_path)

**If you are running from notebook you'll want to comment out the next cell! It's here so that running nbdev tests doesn't take too long, as they test running the entire notebook. It won't be exported so running as a script will be fine**

test == True
constraints = constraints.iloc[:10]  # so tests don't take too long

constraints.head()

# | export

pandarallel.initialize(nb_workers=no_cores, progress_bar=False)
scan_solution = constraints.parallel_apply(
    mmon_gcm.constraintscan.solve_model_for_constraint_scan, args=([model, light_colour]), axis=1)

# | export

solution_df = pd.DataFrame()
for i, solution in scan_solution.items():
    # this is to catch infeasible solutions
    try:
        solution_df = solution_df.append(solution.to_frame().loc[:, "fluxes"], ignore_index=True)
    except:
        solution_df = solution_df.append(pd.Series(dtype=float), ignore_index=True)

solution_df.to_csv(results_path)
