==> src/4.2_run_alternative_flux_modes.ipynb <==
---
output-file: runalternativemodes.html
skip_showdoc: true
---



# Run alternative flux modes simulations

# | default_exp runalternativemodes

As with runconstraintscan.py, this notebook can either be run as a notebook, or using nbdev a python script is produced that can be run in the command line for easier running on e.g. a cluster. It is a notebook in the nbs directory that is symlinked in the nbs/functions directory in order to be exported as a .py file in the mmon-gcm library.

NOTE: I have had issues where for an unknown reason the code hangs while solving. In this case I can check the temporary files to see which solutions are missing and run it again to just get those. For the paper, the blue light, unconstrained, KO solution number 348 hangs with a tolerance of 1e-8 and says infeasible for a tolerance of 1e-7. So the solution list is constructed from combining the remaining temporary files rather than all in one go. 

# | export

import shutil
import sys
from pathlib import Path

import cobra
# from x import y syntax doesn't work because of nbdev export format
import mmon_gcm.alternativemodes
import mmon_gcm.buildingediting
import mmon_gcm.supermodel
import pandas as pd
from pandarallel import pandarallel

# This cell isn't exported to the .py file, so define here if running in notebook rather than as .py on e.g.a cluster

sys.argv = ["script_name",
            "../outputs/alternative_weighting/test_solution.csv",
            "../models/4_stage_GC.json",
            "../outputs/alternative_weighting/alternative_weights.csv",
            "../inputs/arabidopsis_parameters.csv",
            "blue",
            "False",
            "True",
            "2"]

# | export

results_path = sys.argv[1]
model_path = sys.argv[2]
weightings_csv = sys.argv[3]
parameters_csv = sys.argv[4]
light_colour = sys.argv[5]
atpase_constraint = sys.argv[6]
starch_knockout = sys.argv[7]
no_cores = int(sys.argv[8])

# | export

if light_colour != "blue" and light_colour != "white" and light_colour != "nops":
    raise ValueError(f"Please specify either 'blue' or 'white' or 'nops' for light, not {light_colour}")

if atpase_constraint == "True":
    atpase_constraint = True
elif atpase_constraint == "False":
    atpase_constraint = False
else:
    raise ValueError(f"Please specify True or False for the ATPase constraint, not {atpase_constraint}")

if starch_knockout == "True":
    starch_knockout = True
elif starch_knockout == "False":
    starch_knockout = False
else:
    raise ValueError(f"Please specify True or False for the starck knockout, not {starch_knockout}")

# | export

model = cobra.io.load_json_model(model_path)
print("Model imported")

# | export

print(model.solver.configuration.tolerances.integrality)
print(model.solver.configuration.tolerances.feasibility)
model.solver.configuration.tolerances.feasibility = 1e-8  # 1e-9 takes a long time to solve
print(model.solver.configuration.tolerances.feasibility)

# | export

parameters_df = pd.read_csv(parameters_csv, index_col=0)

parameters_df

# | export

arabidopsis_supermodel = mmon_gcm.supermodel.SuperModel(parameters_df.loc[:, "Value"], fba_model=model)
arabidopsis_supermodel.constrain_osmolarity(printouts=False)
arabidopsis_supermodel.constrain_photons(150, printouts=False)
arabidopsis_supermodel.add_maintenance();

## Constrain light, ATPase, starch

# | export

if light_colour == "blue":
    arabidopsis_supermodel.fba_model.reactions.Photon_tx_gc_2.upper_bound = 0
    arabidopsis_supermodel.fba_model.reactions.Photon_tx_me_2.upper_bound = 0
    print("Model constrained with blue light")
elif light_colour == "nops":
    mmon_gcm.buildingediting.set_bounds_multi(arabidopsis_supermodel.fba_model, "Photon_tx_gc", 0, 0)
    print("Photosynthesis prevented in guard cell")
else:
    print("Model constrained with white light")

if atpase_constraint == True:
    gc_atpase_upper_bound = arabidopsis_supermodel.get_atpase_constraint_value(7.48)
    mmon_gcm.buildingediting.set_bounds_multi(arabidopsis_supermodel.fba_model,
                                              "PROTON_ATPase_c_gc", 0, gc_atpase_upper_bound)
    print("Model ATPase constrained")
else:
    print("Model ATPase left unconstrained")

if starch_knockout == True:
    mmon_gcm.buildingediting.set_bounds_multi(arabidopsis_supermodel.fba_model, "RXN_1827_p_gc", 0, 0)
    print("Model starch knocked out")
else:
    print("Model starch left unconstrained")

print("Supermodel established and model constrained")

# | export
weightings = pd.read_csv(weightings_csv, index_col=[0], header=[0])
print("Weightings file imported")

**If you are running from notebook you'll want to comment out the next cell! It's here so that running nbdev tests doesn't take too long, as they test running the entire notebook. It won't be exported so running as a script will be fine**

weightings = weightings.iloc[0:4]  # so tests don't take too long
weightings.head()

# | export

temp_results = results_path.split(".")[0]+"_tmp/"
print(f"Creating temp directory {temp_results} to store solutions as they come in")
Path(temp_results).mkdir(parents=True, exist_ok=True)

# | export

pandarallel.initialize(nb_workers=no_cores, progress_bar=False)
print(f"Solving model for {len(weightings.index)} alternative weightings")
weightings_solution = weightings.parallel_apply(
    mmon_gcm.alternativemodes.solve_model_with_weightings, args=([arabidopsis_supermodel.fba_model, temp_results]), axis=1)

# | export

weightings_solution.to_csv(results_path)
if len(weightings_solution) == len(weightings):
    print(f"All solutions saved to {results_path}")
    print(f"Deleting temp directory {temp_results}")
    shutil.rmtree(temp_results)
