==> src/5_prep_plant_cell.ipynb <==
# Preparing data for 'The Plant Cell' submission

This is pretty slow as the files are large

import os
import cobra

import pandas as pd

i = 1

model = cobra.io.read_sbml_model("../models/4_stage_GC.xml")
cobra.io.write_sbml_model(model, f"../plant_cell_paper/tpc00001-SupplementalDS{i}.xml")
i += 1

model_solutions_legend = "Apart from this one, each sheet contains the solution to the scenario given by the sheet name in the format [light_scenario]_[atpase constrained or unconstrained]_[starch is wt or ko]. Rows are reaction ids, fluxes column is the solution, and minimum and maximum give the flux bounds from FVA for the reactions which it was carried out on."

with pd.ExcelWriter(f"../plant_cell_paper/tpc00001-SupplementalDS{i}.xlsx") as writer:
    legend_df = pd.DataFrame([model_solutions_legend])
    legend_df.to_excel(writer, sheet_name="Legend",index=False, header=False)
    for file in os.listdir("../outputs/model_solutions"):
        filename = os.fsdecode(file)
        if filename.endswith(".csv"):
            solution_df = pd.read_csv("../outputs/model_solutions/" + file, index_col=0)
            solution_df.to_excel(writer, sheet_name=filename[:-4])
i += 1

file_list = ["../outputs/constraint_scan/constraints_df.csv", "../outputs/constraint_scan/constraint_scan_results_blue.csv",
             "../outputs/constraint_scan/constraint_scan_results_white.csv", "../outputs/alternative_weighting/alternative_weights.csv"]
constraints_df_legend = "The data sheet contains the list of constraints used to constrain the model in the constraint scan"
scan_results_blue_legend = "The data sheet contains the flux results for each set of constraints in the constraint scan for the model solved under blue light. The rows correspond to the constraints list and the columns to reaction IDs."
scan_results_white_legend = "The data sheet contains the flux results for each set of constraints in the constraint scan for the model solved under white light. The rows correspond to the constraints list and the columns to reaction IDs."
alternative_weights_legend = "The data sheet contains the list of alternative weights used in the analysis, with the reaction IDs as columns and rows the different sets of weights"
legend_list = [constraints_df_legend, scan_results_blue_legend, scan_results_white_legend, alternative_weights_legend]

for file, legend in zip(file_list, legend_list):
    with pd.ExcelWriter(f"../plant_cell_paper/tpc00001-SupplementalDS{i}.xlsx") as writer:
        legend_df = pd.DataFrame([legend])
        legend_df.to_excel(writer, sheet_name="Legend",index=False, header=False)
        file_df = pd.read_csv(file, index_col=0)
        file_df.to_excel(writer, sheet_name="Data")
        i += 1

for light in ["blue", "white", "nops"]:
    for atpase in ["unconstrained", "constrained"]:
        for starch in ["wt", "ko"]:
            filename = f"../outputs/alternative_weighting/solution_{light}_{atpase}_{starch}.csv"
            legend = f"The data sheet contains the results for solving the alternative weightings for the scenario with {light} light, {atpase} ATPase, and {starch} starch. The columns are reaction IDs and the rows match the alternative weights given in Supplemental Data Set 6"
            try:
                with pd.ExcelWriter(f"../plant_cell_paper/tpc00001-SupplementalDS{i}.xlsx") as writer:
                    legend_df = pd.DataFrame([legend])
                    legend_df.to_excel(writer, sheet_name="Legend",index=False, header=False)
                    solution_df = pd.read_csv(filename, index_col=0)
                    solution_df.to_excel(writer, sheet_name="Data")

                i += 1

            except:
                print(f"No solution for {light}, {atpase}, {starch}")
